+++ src/OpenTelemetry.AutoInstrumentation/Configuration/PluginsConfigurationHelper.cs
@@ -16,6 +16,7 @@
 
 using OpenTelemetry.AutoInstrumentation.Plugins;
 using OpenTelemetry.Metrics;
+using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 
 namespace OpenTelemetry.AutoInstrumentation.Configuration;
@@ -30,5 +31,17 @@ internal static class PluginsConfigurationHelper
     public static MeterProviderBuilder InvokePlugins(this MeterProviderBuilder builder, PluginManager pluginManager)
     {
         return pluginManager.ConfigureMeterProviderBuilder(builder);
+    }
+
+    public static ResourceBuilder InvokePlugins(this ResourceBuilder builder, PluginManager? pluginManager)
+    {
+        if (pluginManager == null)
+        {
+            return builder;
+        }
+        else
+        {
+            return pluginManager.ConfigureResourceBuilder(builder);
+        }
     }
 }